backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-pymongo==3.0.1
python-dotenv==1.1.0
pymongo==4.12.0
gunicorn==23.0.0
openai==1.75.0
celery==5.5.1
dnspython==2.7.0 